Exercise 4.21 Bloglist Expansion: Change the delete blog operation so that a blog can be deleted only by the user who added it. Therefore, deleting a blog is possible only if the token sent with the request is the same as that of the blog's creator.
If deleting a blog is attempted without a token or by an invalid user, the operation returns status code 401 Unauthorized

If deleting a blog is attempted which does not exist, the operation returns status code 404 Not Found

If deleting a blog is attempted by a user which did not add it, the operation returns status 403 Forbidden

+++ controllers/blogs.js
@@ -4,11 +4,13 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 
 
+// GET
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
+// POST
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
@@ -43,6 +45,7 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
+// PUT
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
 
@@ -57,7 +60,27 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(updatedBlog)
 })
 
+// DELETE
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token // Extracted from the tokenExtractor middleware
+
+  if (!token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: 'unauthorized access to delete this blog' })
+  }
+  // If the token matches, delete the blog
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
